move to streaming.read_gbq_table, add logger
